Separated Admin Controller into a generic and extendable Base controller. Added Autoload for properly loading classes
