+++ classes/themes.php
@@ -1,5 +1,5 @@
 <?php
-namespace Grav\Plugin;
+namespace Grav\Plugin\Admin;
 
 /**
  * Admin theme object
