+++ classes/utils.php
@@ -14,6 +14,8 @@ class Utils
     /**
      * Matches an email to a user
      *
+     * @param $email
+     *
      * @return User
      */
     public static function findUserByEmail($email)
