+++ twig/AdminTwigExtension.php
@@ -1,5 +1,5 @@
 <?php
-namespace Grav\Plugin;
+namespace Grav\Plugin\Admin;
 
 use Grav\Common\Grav;
 use Grav\Common\Language\Language;
